Remove lodash as a dependency (#115)
* Remove lodash as a dependency

Apart from making the lib less complex, it also makes it easier to handle with SystemJS. This is important given that the Angular-CLI has embraced SystemJS instead of our usual webpack toolchain.

Also rejigged the deps a bit:
- typings should be a devdepedency
- Angular and RxJS should be peer deps

Connected to #113

* Changed rxjs imports to pull in the parts we need.

This is needed to match the default way Angular-CLI sets up the SystemJS import mappings for RxJs.
Arguably this is the right way to do it anyway; we already switched to this in the v3.x branch.