Convenience wrapper for the dev tools extension. (#155)
* Convenience wrapper for the dev tools extension.

RC4 seems to require the use of AppRef.tick() again to propagate store changes from the dev tools extension.

Since doing it directly in NgRedux is problematic, I've created a wrapper component for it that triggers only for external events (not for actions dispatched in the Angular 2 zone.

This wrapper also neatly hides some of the messiness of accessing `window.devToolsExtension` in TypeScript.

* linter fix

* typo