[Fixes #372] Add support for redux-batch (#397)
Redux-batch (@manaflair/redux-batch) works by overriding Redux's subscribe and notify mechanism. However, the way we were creating an observable out of the store was pushing updates independently of Subscribe. This has the effect that observable.next() was being called independently of redux-batches patched subscribe/notify mechanism and selectors fire on each change of the batched action set.

The fix is to re-create our observable in terms of store.subscribe() in order to catch any adjustments made by 3rd-party enhancers.